+++ gridkit/hex_grid.py
@@ -7,6 +7,7 @@ from pyproj import CRS, Transformer
 from gridkit.base_grid import BaseGrid
 from gridkit.bounded_grid import BoundedGrid
 from gridkit.errors import AlignmentError, IntersectionError
+from gridkit.gridkit_rs import interp
 from gridkit.index import GridIndex, validate_index
 from gridkit.rect_grid import RectGrid
 
@@ -860,35 +861,6 @@ class BoundedHexGrid(BoundedGrid, HexGrid):
         --------
         :py:meth:`.RectGrid.cell_at_point`
         """
-
-        # FIXME: speed up (numba)
-        def get_weight(point, p1, p2, p3):
-            def _project(point, line_points):
-                """Project 'point' onto a line drawn between 'line_points[0]' and 'line_points[1]'
-                Credits to https://stackoverflow.com/a/61343727/22128453
-                """
-                # distance between line_points[0] and line_points[1]
-                line_length = numpy.sum((line_points[0] - line_points[1]) ** 2)
-
-                # project point on line extension connecting line_points[0] and line_points[1]
-                t = (
-                    numpy.sum(
-                        (point - line_points[0]) * (line_points[1] - line_points[0])
-                    )
-                    / line_length
-                )
-
-                return line_points[0] + t * (line_points[1] - line_points[0])
-
-            side_length = numpy.linalg.norm(p1 - p2)
-            dd = p2 + (p2 - p3) / 2
-            if numpy.linalg.norm(dd - p1) > side_length:
-                dd = p2 - (p2 - p3) / 2
-            ad = dd - p1
-
-            projected = _project(point - p1, [p3 - p1, p2 - p1])
-            return numpy.linalg.norm((projected - (point - p1)) / numpy.linalg.norm(ad))
-
         if not isinstance(sample_points, numpy.ndarray):
             sample_points = numpy.array(sample_points)
         original_shape = sample_points.shape
@@ -897,21 +869,11 @@ class BoundedHexGrid(BoundedGrid, HexGrid):
         all_nearby_cells = self.cells_near_point(
             sample_points
         )  # (points, nearby_cells, xy)
-        values = numpy.empty(len(sample_points))
-        for idx, (point, nearby_cells) in enumerate(
-            zip(sample_points, all_nearby_cells)
-        ):
-            nearby_centroids = self.centroid(nearby_cells)
-            p1, p2, p3 = nearby_centroids
-            weights = numpy.array(
-                [
-                    get_weight(point, p1, p2, p3),
-                    get_weight(point, p2, p1, p3),
-                    get_weight(point, p3, p2, p1),
-                ]
-            )
-            values[idx] = numpy.sum(weights * self.value(nearby_cells))
-
+        nearby_centroids = self.centroid(all_nearby_cells)
+        weights = interp.linear_interp_weights_triangles(
+            sample_points, nearby_centroids
+        )
+        values = numpy.sum(weights * self.value(all_nearby_cells), axis=1)
         # TODO: remove rows and cols with nans around the edge after bilinear
         return values.reshape(*original_shape[:-1])